+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsetypescript.py
@@ -42,9 +42,9 @@ class EvaluatorResponseTypescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-08-20T04:17:08.353Z"
+    created: Optional[str] = "2026-08-20T11:48:22.539Z"
 
-    updated: Optional[str] = "2026-08-20T04:17:08.353Z"
+    updated: Optional[str] = "2026-08-20T11:48:22.539Z"
 
     updated_by_id: OptionalNullable[str] = UNSET