+++ packages/orq-rc/src/orq_ai_sdk/models/retrieveannotationqueueop.py
@@ -94,7 +94,7 @@ class RetrieveAnnotationQueueResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-20T04:17:06.109Z")
+    updated: Optional[datetime] = parse_datetime("2026-08-20T11:48:20.273Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")