+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -2068,7 +2068,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01M0EP6HXS5QZKZ6PPCSKACZ99"
+    id: Optional[str] = "01M0FG0STP7J1D63SR3NQKW20C"
 
     description: Optional[str] = None